Parser optimisation.
In particualr this input ran painfully slow:
(((((1)))))
Now we parse it instantly. On the binary operator style rules for
instance on addsub the implementation was :

addsub: divmul Some([+-] divmul) | divmul

This style goes very deep on the lhs before it realises that it fails on
the operator, and parses with a single divmul instead. Also wastfully
throws away the already parsed divmul. Now we have c.Any which is a
repetition any number of times:

addsub: divmul Any([+-] divmul)