Reimplemented variable lookup
This is primarily a performance improvement. 44% speedup on
examples/euler_35.calc.

There is a new stage in the interpreter after parse and before
evaluation, called STRewrite. STRewrite replaces local variable names
with int indices into the stack frame of the function call.

Introduced the 3 variable types - based on lexical scope.

All examples from before work the same.

!!!WARNING!!!

This is a breaking change tough, changing language semantics. Previously
variables were looked up in the call stack from last stack frame to
first. That meant that a variable visible in the caller could be
accessed within a function call. This is not true any more.