make errors contain string pointers.
Small performance improvement. Funnily enough the code

{
  if cond something() else 0
  something()
}

runs faster than without else 0 if cond fails often. That is because the
if creates a NoResultError which is then thrown away in the block.
Creating the error shouldn't have been a costly operation as errors are
not mutable, so a statically allocated error should have been fine.
However we still allocated a new string every time the error was
created.
Now the built in errors point to the same strings.