This is an implementation of iterators
This is currently not performant enough, because it simulates coroutines
with 2 go routines that switch lock-stepped with two channels. The
switching overhead is too much for this to be feasible for iterators.

Probably we would need a bytecode interpreter first, in which we can
have fine grain control over context switches. This needs to be close to
a simple jump with minimal overhead for switching between stacks.