bytecode interpreter (#6)
Implemented bytecode evaluation, vm. 64 bit instructions. static data is in DS and multiple occurances of literal values emit multiple DS allocations. This could be optimised with a hashmap for space, but would inccur runtime penalty.

There can be a lot optimised in the size of the CS, however it would hardly give a performance difference as the instructions executed would be the same-ish just less of them.. ie, compiling an if or a while could use some of the prelude from a containing if or while

Deleted AST based evaluator. Gives a small performance improvement on euler_35. There can be more things optimised now that we have finer control over execution.

exit didn't exit if it wasn't given a number now it always exits.