Removed error as a value type. (#9)
Errors as a value type was a bad design idea. it avoids runtime errors, at the cost of obscuring problems allowing errors to slip far from where they occured etc. In the new design when there is a runtime error the interpreter will halt and produce backtrace and debug info.

There is the problem of what happens with

```
f  = () -> if false 1
x = f()
```

as not having errors doesn't allow us for x to take a value. We will have nil for this case. nil will be fine as long as it's ignored. The call to f will succeed. The assigment to x will become runtime error, because it won't be allowed to do anything with nil. Similar example will be the uninitialised local.

```
f  = () -> {
  if false x = 1
  x
}
```

at this point again nil is returned from f which will be fine as long as it's ignored in the caller.

nil won't be exposed as a value. It's just an internal type that a value can take, so we can return it as a result of functions , control structures (like a loop that doesn't run it's body), etc.

when a value of nil is then used in an expression it will cause runtime error.