Optimisation for binary / unary operator expressions (#19)
Introduced a tmp register in the vm. Introduced temp versions of instructions. A temp instruction takes one less src than the normal counterpart, and sources the temp register for the missing operand and targets the temp register for destination. This avoids a lot of push/pop on the stack which is supposed to be faster as the tmp lives inside the run loop as local variable.

Basically this

        1+2+3+4+5
        69 | 0X08360000000D000E : ADD DS[13] DS[14]
        70 | 0X082E00000000000F : ADD STCK DS[15]
        71 | 0X082E000000000010 : ADD STCK DS[16]
        72 | 0X082E000000000011 : ADD STCK DS[17]
        
turns into 

        69 | 0X063700000000000D : MOV TMP DS[13]
        70 | 0X880700000000000E : ADDTMP DS[14]
        71 | 0X880700000000000F : ADDTMP DS[15]
        72 | 0X8807000000000010 : ADDTMP DS[16]
        73 | 0X8807000000000011 : ADDTMP DS[17]
        74 | 0X8200000000000000 : PUSHTMP
        
It's more instructions, but the former does 4 pushes and 3 pops, the latter does 1 push and an extra mov

Gives a marginal speed improvement on euler_35