while loop optimisation
When a while loop isn't discarding, we eliminated a pair of push/pop
from the loop body assuming the loop body didn't put its result on the
stack. This is by turning the loop into a post check loop where the post
check jump doesn't jump on the initial pop but jumps directly back on
the body. We need to add an extra push at the end though. This is under
the assumption that the condition code cannot corrupt the body results,
which is safe to assume as it's an expression not a statement.