small optimisation in yield (#55)
yield most of the time doesn't need to push in the context it's invoked
at. We often compiled in an extra push. However sometimes it does need
to leave the yielded value on the stack. We can use the tmp register to
save the value, and needed we can do a PUSHTMP. This is safe because if
yield was invoked from a for loop the for will throw away the iteration
result anyway. And if yield wasn't called from a for loop, there is no
context switch that would corrupt tmp.

small speed improvement on euler_35_new